requirements-dev.txt
dj_database_url==0.5.0
dj_static==0.0.6
Django==3.1.1
djangorestframework==3.13.1
Markdown==3.2.2
pandas==1.1.5
python-decouple==3.6
django-markdownify
django-widget-tweaks
django-cors-headers
django_storages==1.13.1